+++ source/DgtAngel.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.0.31612.314
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{DC5E410E-067B-48BA-8C57-F231E35A8790}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DgtEbDllWrapper", "DgtEbDllWrapper\DgtEbDllWrapper.csproj", "{B9F97188-1996-4F0F-BD67-734EA3F672D1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DgtRabbitWrapper", "DgtEbDllWrapper\DgtRabbitWrapper.csproj", "{B9F97188-1996-4F0F-BD67-734EA3F672D1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DgtCherub", "DgtCherub\DgtCherub.csproj", "{A3126BB7-0C51-43CA-9075-2863CE97E2F6}"
 EndProject
@@ -50,7 +50,7 @@ Project("{C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5}") = "DgtCherubPackaging", "DgtCh
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "WingedHorseBuild", "WingedHorseBuild", "{E3242AD1-F11C-46F0-A273-3D38440CB36C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChessHelpers", "ChessHelpers\ChessHelpers.csproj", "{D891B5C2-11B5-44BC-A261-D6534BEEBDBE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChessHelpers", "ChessHelpers\ChessHelpers.csproj", "{D891B5C2-11B5-44BC-A261-D6534BEEBDBE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
