+++ source/DgtCherub/DgtCherub.csproj
@@ -122,7 +122,7 @@
   <ItemGroup>
     <ProjectReference Include="..\ChessHelpers\ChessHelpers.csproj" />
     <ProjectReference Include="..\DgtAngelShared\DgtAngelShared.csproj" />
-    <ProjectReference Include="..\DgtEbDllWrapper\DgtEbDllWrapper.csproj" />
+    <ProjectReference Include="..\DgtEbDllWrapper\DgtRabbitWrapper.csproj" />
     <ProjectReference Include="..\DgtLiveChessWrapper\DgtLiveChessWrapper.csproj" />
     <ProjectReference Include="..\DynamicBoard\DynamicBoard.csproj" />
   </ItemGroup>
