+++ source/DgtEbDllWrapper/DgtRabbitWrapper.csproj
